+++ front/src/components/pages/mypage/sentences/index.vue
