+++ front/src/router/index.ts
@@ -11,8 +11,9 @@ const Sentence = () => import("@/components/pages/sentences/show.vue")
 const EditSentence = () => import("@/components/pages/sentences/edit.vue")
 const BookmarkSentences = () => import("@/components/pages/sentences/bookmark.vue")
 const Profile = () => import("@/components/pages/profile/index.vue")
-const Mypage = () => import("@/components/pages/mypage/index.vue")
-const MySentences = () => import("@/components/pages/mypage/sentences.vue")
+const MyHistory = () => import("@/components/pages/mypage/history/index.vue")
+const MyRecord = () => import("@/components/pages/mypage/record/index.vue")
+const MySentences = () => import("@/components/pages/mypage/sentences/index.vue")
 const History = () => import("@/components/pages/history/show.vue")
 const Privacy = () => import("@/components/pages/static/Privacy.vue")
 const Terms = () => import("@/components/pages/static/Terms.vue")
@@ -72,9 +73,15 @@ const routes: Array<RouteRecordRaw> = [
     meta: { requiresAuth: true }
   },
   {
-    path: "/mypage",
-    name: "Mypage",
-    component: Mypage,
+    path: "/mypage/history",
+    name: "MyHistory",
+    component: MyHistory,
+    meta: { requiresAuth: true }
+  },
+  {
+    path: "/mypage/record",
+    name: "MyRecord",
+    component: MyRecord,
     meta: { requiresAuth: true }
   },
   {
